Addressed reviewer comments: Included functions on export_widget class,  reverted changes on imports and ui object
